+++ phypno/ioeeg/__init__.py
@@ -10,4 +10,5 @@ from .mnefiff import write_mnefiff
 from .fieldtrip import FieldTrip, write_fieldtrip
 from .phypno import Phypno, write_phypno
 from .ieeg_org import IEEG_org
+from .opbox import OpBox
 # from .eeglab import Eeglab, write_eeglab