+++ test/test_phypno_detect_ripple.py
@@ -1,8 +1,0 @@
-from phypno.graphoelement import Ripple
-from phypno.detect import DetectRipple
-
-
-def test_ripple_01():
-    det_rp = DetectRipple()
-    rp = det_rp(None)
-    assert isinstance(rp, Ripple)