+++ docs/source/changelog.rst
@@ -1,5 +1,12 @@
 Change Log
 ==========
+Version 17
+----------
+- **17.6**: Use correct start time for opbox
+- **17.4**: Upload correct version to pypi and docs
+- **17.1**: Self-host documentation
+- **17.0**: Read OpBox files
+
 Version 16
-------------
+----------
 - **16.9**: Added some documentation for installation